Update action.yml
`-x` ensures that the entire line is matched exactly.

For example, if a user in `allowed_users` is `Sabit` and someone created an account with the username `abit`, the current implementation could mistakenly allow `abit` because it partially matches `Sabit`. Using the `-x` option ensures that the entire line is matched exactly, preventing partial matches from succeeding.